Update Maven and plugin versions
The Maven and various plugin versions in the pom.xml file have been updated. Also, the Maven distribution URL in the maven-wrapper.properties file has been updated to reflect the new Maven version. These updates ensure that the project is built using the latest stable versions of Maven and the plugins.